+++ src/olmo_core/internal/common.py
@@ -2,6 +2,7 @@ import logging
 from functools import lru_cache
 from typing import List, Optional
 
+import torch
 from beaker import Beaker, BeakerError
 
 from olmo_core.io import is_url
@@ -36,9 +37,9 @@ def get_beaker_username() -> Optional[str]:
 
 def get_root_dir(cluster: str) -> str:
     root_dir: str = "weka://oe-training-default/ai2-llm"
-    if "jupiter" in cluster:
+    if "cirrascale" in cluster:
         root_dir = "/weka/oe-training-default/ai2-llm"
-    elif "augusta" in cluster:
+    elif "google" in cluster:
         root_dir = "gs://ai2-llm"
     elif "local" in cluster:
         root_dir = "gs://ai2-llm"
@@ -126,12 +127,15 @@ CLUSTER_TO_GPU_TYPE = {
     "ai2/jupiter-cirrascale-2": "NVIDIA H100 80GB HBM3",
     "ai2/pluto-cirrascale": "NVIDIA H100",
     "ai2/augusta-google-1": "NVIDIA H100",
+    "ai2/titan-cirrascale": "NVIDIA B200",
 }
 
 
 def get_gpu_type(cluster: str) -> str:
     if cluster in CLUSTER_TO_GPU_TYPE:
         return CLUSTER_TO_GPU_TYPE[cluster]
+    elif cluster == "local":
+        return torch.get_default_device().type
     else:
         log.warning(f"Missing cluster '{cluster}' in CLUSTER_TO_GPU_TYPE mapping")
         beaker = get_beaker_client()