+++ src/olmo_core/internal/model_ladder.py
@@ -18,7 +18,7 @@ from olmo_core.train import (
 )
 from olmo_core.train.callbacks import CometCallback, ConfigSaverCallback, WandBCallback
 from olmo_core.train.train_module import TransformerTrainModuleConfig
-from olmo_core.utils import prepare_cli_environment, seed_all
+from olmo_core.utils import get_default_device, prepare_cli_environment, seed_all
 
 from .common import build_launch_config, get_gpu_type, get_root_dir
 
@@ -39,6 +39,7 @@ class LadderRunConfig(Config):
 class SubCmd(StrEnum):
     launch = "launch"
     train = "train"
+    train_single = "train_single"
     dry_run = "dry_run"
 
     def prepare_environment(self):
@@ -46,6 +47,8 @@ class SubCmd(StrEnum):
             prepare_cli_environment()
         elif self == SubCmd.train:
             prepare_training_environment()
+        elif self == SubCmd.train_single:
+            prepare_training_environment(backend=None)
         else:
             raise NotImplementedError(self)
 
@@ -57,13 +60,27 @@ class SubCmd(StrEnum):
             config.launch.launch(follow=True)
         elif self == SubCmd.dry_run:
             pass
-        elif self == SubCmd.train:
+        elif self in (SubCmd.train, SubCmd.train_single):
+            if self == SubCmd.train_single:
+                for parallelism_strategy in {"d", "t", "e", "c"}:
+                    setting_name = f"{parallelism_strategy}p_config"
+                    if getattr(config.train_module, setting_name) is not None:
+                        log.warning(
+                            "%s is set to %s, but you can't use data parallelism when running on a single node. Disabling.",
+                            setting_name,
+                            getattr(config.train_module, setting_name),
+                        )
+                        setattr(config.train_module, setting_name, None)
+                init_device = str(get_default_device())
+            else:
+                init_device = "meta"
+
             try:
                 # Set RNG states on all devices.
                 seed_all(config.ladder.init_seed)
 
                 # Build components.
-                model = config.model.build(init_device="meta")
+                model = config.model.build(init_device=init_device)
                 train_module = config.train_module.build(model)
                 dataset = config.dataset.build()
                 data_loader = config.data_loader.build(
@@ -133,10 +150,11 @@ def main(ladder_builder: Callable[[str], ModelLadder]):
 [yellow]Usage:[/] [i blue]python[/] [i cyan]{sys.argv[0]}[/] [i b magenta]{'|'.join(SubCmd)}[/] [i b]SIZE RUN_DURATION CLUSTER[/] [i][OVERRIDES...][/]
 
 [b]Subcommands[/]
-[b magenta]launch:[/]      Launch the script on Beaker with the [b magenta]train[/] subcommand.
-[b magenta]train:[/]       Run the trainer. You usually shouldn't invoke the script with this subcommand directly.
-             Instead use [b magenta]launch[/] or run it with torchrun.
-[b magenta]dry_run:[/]     Pretty print the config to run and exit.
+[b magenta]launch:[/]       Launch the script on Beaker with the [b magenta]train[/] subcommand.
+[b magenta]train:[/]        Run the trainer. You usually shouldn't invoke the script with this subcommand directly.
+              Instead use [b magenta]launch[/] or run it with torchrun.
+[b magenta]train_single:[/] Run the trainer on a single device (GPU, CPU, MPS). num_nodes is ignored.
+[b magenta]dry_run:[/]      Pretty print the config to run and exit.
 
 [b]Examples[/]
 $ [i]python {sys.argv[0]} {SubCmd.launch} 1B 1xC ai2/pluto-cirrascale --launch.num_nodes=2[/]