+++ src/olmo_core/optim/scheduler.py
@@ -50,14 +50,84 @@ class ConstantWithWarmup(Scheduler):
     Constant learning rate schedule with a warmup.
     """
 
-    warmup_steps: int = 2000
+    warmup_steps: Optional[int] = 2000
+    warmup_fraction: Optional[float] = None
     warmup_min_lr: float = 0.0
+
+    def __post_init__(self):
+        if (self.warmup_fraction is None) == (self.warmup_steps is None):
+            raise OLMoConfigurationError(
+                "Either warmup_fraction or warmup_steps must be specified."
+            )
+        if self.warmup_fraction is not None and (
+            self.warmup_fraction < 0 or self.warmup_fraction > 1
+        ):
+            raise OLMoConfigurationError("warmup_fraction must be between 0 and 1.")
 
     def get_lr(
         self, initial_lr: Union[float, torch.Tensor], step: int, max_steps: int
     ) -> Union[float, torch.Tensor]:
-        if step <= self.warmup_steps:
-            return _linear_warmup(initial_lr, step, self.warmup_steps, self.warmup_min_lr)
+        if self.warmup_steps is None:
+            assert self.warmup_fraction is not None
+            warmup_steps = round(max_steps * self.warmup_fraction)
+        else:
+            warmup_steps = self.warmup_steps
+
+        if step <= warmup_steps:
+            return _linear_warmup(initial_lr, step, warmup_steps, self.warmup_min_lr)
+
+        del step, max_steps
+        return initial_lr
+
+
+@dataclass
+class WSD(Scheduler):
+    """
+    Warmup-stable-decay scheduler
+    """
+
+    warmup_steps: Optional[int] = 2000
+    warmup_fraction: Optional[float] = None
+    decay_steps: Optional[int] = None
+    decay_fraction: Optional[float] = 0.1
+    warmup_min_lr: float = 0.0
+    decay_min_lr: float = 0.0
+
+    def __post_init__(self):
+        if (self.warmup_fraction is None) == (self.warmup_steps is None):
+            raise OLMoConfigurationError(
+                "Either warmup_fraction or warmup_steps must be specified."
+            )
+        if self.warmup_fraction is not None and (
+            self.warmup_fraction < 0 or self.warmup_fraction > 1
+        ):
+            raise OLMoConfigurationError("warmup_fraction must be between 0 and 1.")
+
+        if (self.decay_fraction is None) == (self.decay_steps is None):
+            raise OLMoConfigurationError("Either decay_fraction or decay_steps must be specified.")
+        if self.decay_fraction is not None and (self.decay_fraction < 0 or self.decay_fraction > 1):
+            raise OLMoConfigurationError("decay_fraction must be between 0 and 1.")
+
+    def get_lr(
+        self, initial_lr: Union[float, torch.Tensor], step: int, max_steps: int
+    ) -> Union[float, torch.Tensor]:
+        if self.warmup_steps is None:
+            assert self.warmup_fraction is not None
+            warmup_steps = round(max_steps * self.warmup_fraction)
+        else:
+            warmup_steps = self.warmup_steps
+
+        if step <= warmup_steps:
+            return _linear_warmup(initial_lr, step, warmup_steps, self.warmup_min_lr)
+
+        if self.decay_steps is None:
+            assert self.decay_fraction is not None
+            decay_steps = round(max_steps * self.decay_fraction)
+        else:
+            decay_steps = self.decay_steps
+
+        if step >= max_steps - decay_steps:
+            return _linear_decay(initial_lr, max_steps - step, decay_steps, self.decay_min_lr)
 
         del step, max_steps
         return initial_lr
@@ -71,21 +141,39 @@ class LinearWithWarmup(Scheduler):
 
     alpha_f: float = 0.1
     t_max: Optional[int] = None
-    warmup_steps: int = 2000
+    warmup_steps: Optional[int] = 2000
+    warmup_fraction: Optional[float] = None
     warmup_min_lr: float = 0.0
+
+    def __post_init__(self):
+        if (self.warmup_fraction is None) == (self.warmup_steps is None):
+            raise OLMoConfigurationError(
+                "Either warmup_fraction or warmup_steps must be specified."
+            )
+        if self.warmup_fraction is not None and (
+            self.warmup_fraction < 0 or self.warmup_fraction > 1
+        ):
+            raise OLMoConfigurationError("warmup_fraction must be between 0 and 1.")
 
     def get_lr(
         self, initial_lr: Union[float, torch.Tensor], step: int, max_steps: int
     ) -> Union[float, torch.Tensor]:
         max_steps = max_steps if self.t_max is None else self.t_max
         eta_min = initial_lr * self.alpha_f
-        if step < self.warmup_steps:
-            return _linear_warmup(initial_lr, step, self.warmup_steps, self.warmup_min_lr)
+
+        if self.warmup_steps is None:
+            assert self.warmup_fraction is not None
+            warmup_steps = round(max_steps * self.warmup_fraction)
+        else:
+            warmup_steps = self.warmup_steps
+
+        if step < warmup_steps:
+            return _linear_warmup(initial_lr, step, warmup_steps, self.warmup_min_lr)
         elif step >= max_steps:
             return eta_min
         else:
-            step = step - self.warmup_steps
-            max_steps = max_steps - self.warmup_steps
+            step = step - warmup_steps
+            max_steps = max_steps - warmup_steps
             return initial_lr - (initial_lr - eta_min) * (step / max_steps)
 
 
@@ -96,18 +184,35 @@ class InvSqrtWithWarmup(Scheduler):
     """
 
     alpha_f: float = 0.1
-    warmup_steps: int = 2000
+    warmup_steps: Optional[int] = 2000
+    warmup_fraction: Optional[float] = None
     warmup_min_lr: float = 0.0
+
+    def __post_init__(self):
+        if (self.warmup_fraction is None) == (self.warmup_steps is None):
+            raise OLMoConfigurationError(
+                "Either warmup_fraction or warmup_steps must be specified."
+            )
+        if self.warmup_fraction is not None and (
+            self.warmup_fraction < 0 or self.warmup_fraction > 1
+        ):
+            raise OLMoConfigurationError("warmup_fraction must be between 0 and 1.")
 
     def get_lr(
         self, initial_lr: Union[float, torch.Tensor], step: int, max_steps: int
     ) -> Union[float, torch.Tensor]:
-        if step < self.warmup_steps:
-            return _linear_warmup(initial_lr, step, self.warmup_steps, self.warmup_min_lr)
+        if self.warmup_steps is None:
+            assert self.warmup_fraction is not None
+            warmup_steps = round(max_steps * self.warmup_fraction)
+        else:
+            warmup_steps = self.warmup_steps
+
+        if step < warmup_steps:
+            return _linear_warmup(initial_lr, step, warmup_steps, self.warmup_min_lr)
         del max_steps
 
         eta_min = initial_lr * self.alpha_f
-        return eta_min + (initial_lr - eta_min) * sqrt(self.warmup_steps / step)
+        return eta_min + (initial_lr - eta_min) * sqrt(warmup_steps / step)
 
 
 @dataclass
@@ -116,24 +221,74 @@ class CosWithWarmup(Scheduler):
     Cosine learning rate schedule with a warmup.
     """
 
-    warmup_steps: int = 2000
+    warmup_steps: Optional[int] = 2000
+    warmup_fraction: Optional[float] = None
     alpha_f: float = 0.1
     t_max: Optional[int] = None
     warmup_min_lr: float = 0.0
+
+    def __post_init__(self):
+        if (self.warmup_fraction is None) == (self.warmup_steps is None):
+            raise OLMoConfigurationError(
+                "Either warmup_fraction or warmup_steps must be specified."
+            )
+        if self.warmup_fraction is not None and (
+            self.warmup_fraction < 0 or self.warmup_fraction > 1
+        ):
+            raise OLMoConfigurationError("warmup_fraction must be between 0 and 1.")
 
     def get_lr(
         self, initial_lr: Union[float, torch.Tensor], step: int, max_steps: int
     ) -> Union[float, torch.Tensor]:
         max_steps = max_steps if self.t_max is None else self.t_max
         eta_min = initial_lr * self.alpha_f
-        if step < self.warmup_steps:
-            return _linear_warmup(initial_lr, step, self.warmup_steps, self.warmup_min_lr)
+
+        if self.warmup_steps is None:
+            assert self.warmup_fraction is not None
+            warmup_steps = round(max_steps * self.warmup_fraction)
+        else:
+            warmup_steps = self.warmup_steps
+
+        if step < warmup_steps:
+            return _linear_warmup(initial_lr, step, warmup_steps, self.warmup_min_lr)
         elif step >= max_steps:
             return eta_min
         else:
-            step = step - self.warmup_steps
-            max_steps = max_steps - self.warmup_steps
+            step = step - warmup_steps
+            max_steps = max_steps - warmup_steps
             return eta_min + (initial_lr - eta_min) * (1 + cos(pi * step / max_steps)) / 2
+
+
+@dataclass
+class CosWithWarmupAndLinearDecay(CosWithWarmup):
+    """
+    Cosine learning rate schedule with a warmup, cut short at the end and followed by a linear decay.
+    """
+
+    decay_steps: Optional[int] = None
+    decay_fraction: Optional[float] = 0.1
+    decay_min_lr: float = 0.0
+
+    def __post_init__(self):
+        if (self.decay_fraction is None) == (self.decay_steps is None):
+            raise OLMoConfigurationError("Either decay_fraction or decay_steps must be specified.")
+        if self.decay_fraction is not None and (self.decay_fraction < 0 or self.decay_fraction > 1):
+            raise OLMoConfigurationError("decay_fraction must be between 0 and 1.")
+
+    def get_lr(
+        self, initial_lr: Union[float, torch.Tensor], step: int, max_steps: int
+    ) -> Union[float, torch.Tensor]:
+        if self.decay_steps is None:
+            assert self.decay_fraction is not None
+            decay_steps = round(max_steps * self.decay_fraction)
+        else:
+            decay_steps = self.decay_steps
+
+        if step >= max_steps - decay_steps:
+            final_cosine_lr = super().get_lr(initial_lr, max_steps - decay_steps, max_steps)
+            return _linear_decay(final_cosine_lr, max_steps - step, decay_steps, self.decay_min_lr)
+
+        return super().get_lr(initial_lr, step, max_steps)
 
 
 def _linear_warmup(
@@ -142,6 +297,20 @@ def _linear_warmup(
     if isinstance(initial_lr, float):  # not worth the potential host-device sync if it's a tensor
         assert 0 <= warmup_min_lr < initial_lr
     return warmup_min_lr + (initial_lr - warmup_min_lr) * min(step, warmup_steps) / warmup_steps
+
+
+def _linear_decay(
+    initial_lr: Union[float, torch.Tensor],
+    step_from_end: int,
+    decay_steps: int,
+    decay_min_lr: float = 0.0,
+) -> Union[float, torch.Tensor]:
+    if isinstance(initial_lr, float):  # not worth the potential host-device sync if it's a tensor
+        assert 0 <= decay_min_lr < initial_lr
+
+    return (
+        decay_min_lr + (initial_lr - decay_min_lr) * min(step_from_end, decay_steps) / decay_steps
+    )
 
 
 @dataclass